improve tests & some functionalities (#46)
* refactor factory function gate
* refactor matrix form of RotationGate
* improve test coverage
* fix some bugs

+++ test/Register.jl
@@ -9,7 +9,7 @@ using Compat.Test
 
     test_data = zeros(Complex64, 2^5, 3)
     reg = register(test_data, 3)
-    @test typeof(reg) == Register{UInt(3), Complex64}
+    @test typeof(reg) == Register{3, Complex64}
     @test address(reg) == collect(1:5)
     @test nqubit(reg) == 5
     @test nbatch(reg) == 3
@@ -30,7 +30,7 @@ using Compat.Test
     @test state(creg) == state(reg)
     @test address(creg) == address(reg)
     @test state(creg) !== state(reg)
-    @test address(creg) !== address(reg) 
+    @test address(creg) !== address(reg)
 end
 
 @testset "Focus" begin
